+++ tests_system/lobster_json/lobsterjsonasserter.py
@@ -15,7 +15,8 @@ class LobsterJsonAsserter(Asserter):
 
     def assertStdOutNumAndFileDeprecated(self, num_items: int, out_file: str, schema: str, version: int):
         self.assertStdOutText(
-            f"Lobster file version {version} containing 'schema' = '{schema}' is deprecated, "
+            f"Lobster file version {version} "
+            f"containing 'schema' = '{schema}' is deprecated, "
             f"please migrate to version 5\n"
             f"lobster-json: wrote {num_items} items to {out_file}\n"
         )