+++ tests_system/lobster_pkg/lobster_pkg_asserter.py
@@ -12,10 +12,13 @@ class LobsterPkgAsserter(Asserter):
         self.assertStdOutText(
             f"lobster-pkg: wrote {num_items} items to {out_file}\n"
         )
-    
-    def assertStdOutNumAndFileDeprecated(self, num_items: int, out_file: str, schema: str, version: int):
+
+    def assertStdOutNumAndFileDeprecated(
+            self, num_items: int, out_file: str, schema: str, version: int
+        ):
         self.assertStdOutText(
-            f"Lobster file version {version} containing 'schema' = '{schema}' is deprecated, "
+            f"Lobster file version {version} "
+            f"containing 'schema' = '{schema}' is deprecated, "
             f"please migrate to version 5\n"
-            f"Written output for {num_items} items to {out_file}\n"
+            f"lobster-pkg: wrote {num_items} items to {out_file}\n"
         )