+++ tests_system/lobster_python/test_valid_cases.py
@@ -5,7 +5,7 @@ from tests_system.lobster_python.lobster_python_system_test_case_base import (
     LobsterPythonSystemTestCaseBase,
 )
 from tests_system.lobster_python.\
-    lobster_python_asserter import LobsterPythonTestAsserter as Asserter
+    lobster_python_asserter import LobsterPythonAsserter as Asserter
 
 
 class ValidCasesPython(LobsterPythonSystemTestCaseBase):
@@ -16,6 +16,27 @@ class ValidCasesPython(LobsterPythonSystemTestCaseBase):
     def test_valid_python(self):
         """Simple Python file and verifies the generated lobster output."""
         OUT_FILE = "basic.lobster"
+
+        self._test_runner.declare_input_file(self._data_directory / "basic.py")
+
+        self._test_runner.cmd_args.files = ["basic.py"]
+        self._test_runner.cmd_args.single = True
+        self._test_runner.cmd_args.out = OUT_FILE
+        self._test_runner.cmd_args.kind = "imp"
+
+
+        self._test_runner.declare_output_file(self._data_directory / OUT_FILE)
+
+        completed_process = self._test_runner.run_tool_test()
+        asserter = Asserter(self, completed_process, self._test_runner)
+        asserter.assertNoStdErrText()
+        asserter.assertStdOutNumAndFileDeprecated(3, OUT_FILE, "lobster-imp-trace", 3)
+        asserter.assertExitCode(0)
+        asserter.assertOutputFiles()
+
+    def test_valid_python_no_schema(self):
+        """Simple Python file and verifies the generated lobster output."""
+        OUT_FILE = "basic_no_schema.lobster"
 
         self._test_runner.declare_input_file(self._data_directory / "basic.py")