+++ tests_system/lobster_report/test_justification.py
@@ -38,8 +38,10 @@ class ReportJustificationTest(LobsterReportSystemTestCaseBase):
         completed_process = self._test_runner.run_tool_test()
         asserter = Asserter(self, completed_process, self._test_runner)
         asserter.assertNoStdErrText()
-        asserter.assertStdOutText(f"{conf_file}: lobster warning: configuration file format '.conf' "
-                                  "is deprecated, please migrate to '.yaml' format\n")
+        asserter.assertStdOutText(
+            f"{conf_file}: lobster warning: configuration file format '.conf' "
+            f"is deprecated, please migrate to '.yaml' format\n"
+        )
         asserter.assertExitCode(0)
         asserter.assertOutputFiles()
 
@@ -53,14 +55,18 @@ class ReportJustificationTest(LobsterReportSystemTestCaseBase):
         This test checks that the lobster report tool can handle justifications
         and coverage changes according to justifications and generate lobster report.
         """
-        self._test_runner.declare_input_file(self._data_directory /
-                                             "just_policy_no_schema.yaml")
-        self._test_runner.declare_input_file(self._data_directory /
-                                             "just_usecases_no_schema.lobster")
-        self._test_runner.declare_input_file(self._data_directory /
-                                             "just_system_requirements_no_schema.lobster")
-        self._test_runner.declare_input_file(self._data_directory /
-                                             "just_software_requirements_no_schema.lobster")
+        self._test_runner.declare_input_file(
+            self._data_directory / "just_policy_no_schema.yaml"
+        )
+        self._test_runner.declare_input_file(
+            self._data_directory / "just_usecases_no_schema.lobster"
+        )
+        self._test_runner.declare_input_file(
+            self._data_directory / "just_system_requirements_no_schema.lobster"
+        )
+        self._test_runner.declare_input_file(
+            self._data_directory / "just_software_requirements_no_schema.lobster"
+        )
 
         conf_file = "just_policy_no_schema.yaml"
         out_file = "just_report_yaml_no_schema.lobster"