+++ tests_system/lobster_trlc/test_inputs_and_inputs_from_file.py
@@ -47,7 +47,8 @@ class InputFromFilesAndInputsTest(LobsterTrlcSystemTestCaseBase):
         lobster_schema = "lobster-req-trace"
         lobster_version = 4
         asserter.assertStdOutText(
-            f"Lobster file version {lobster_version} containing 'schema' = '{lobster_schema}' is deprecated, "
+            f"Lobster file version {lobster_version} "
+            f"containing 'schema' = '{lobster_schema}' is deprecated, "
             f"please migrate to version 5\n"
             f"lobster-trlc: wrote 2 items to "
             f"{OUT_FILE}\n")