+++ tests_system/lobster_trlc/test_inputs_zero.py
@@ -29,7 +29,8 @@ class ZeroInputTest(LobsterTrlcSystemTestCaseBase):
         lobster_schema = "lobster-req-trace"
         lobster_version = 4
         asserter.assertStdOutText(
-            f"Lobster file version {lobster_version} containing 'schema' = '{lobster_schema}' is deprecated, "
+            f"Lobster file version {lobster_version} "
+            f"containing 'schema' = '{lobster_schema}' is deprecated, "
             f"please migrate to version 5\n"
             f"lobster-trlc: wrote 0 items to "
             f"{OUT_FILE}\n")
@@ -75,7 +76,8 @@ class ZeroInputTest(LobsterTrlcSystemTestCaseBase):
         lobster_schema = "lobster-req-trace"
         lobster_version = 4
         asserter.assertStdOutText(
-            f"Lobster file version {lobster_version} containing 'schema' = '{lobster_schema}' is deprecated, "
+            f"Lobster file version {lobster_version} "
+            f"containing 'schema' = '{lobster_schema}' is deprecated, "
             f"please migrate to version 5\n"
             f"lobster-trlc: wrote 0 items to "
             f"{OUT_FILE}\n")